Move title view to separate file
